💄 polish(agent-topic-manager): lighter bulk-bar shadow, transparent tool-auth alert, preserve sub-route on agent switch
- BulkActionBar: tone down the floating pill shadow from a heavy 24%/16%
  stack to a softer 8%/6% pair so it stops competing with the list rows.
- ToolAuthAlert: drop the secondary-tint fill (`background: transparent`)
  so the panel reads as a calm hint, not a warning. Reword the hint copy
  to "技能未授权或未配置时，相关技能无法使用，可能导致助理能力受限或报错" /
  matching EN.
- Sidebar agent switcher: clicking Lobe AI (Inbox) from `/agent/X/topics`
  now lands on `/agent/inbox/topics` instead of dropping back to the
  default chat URL. Extracts the existing `AgentItem` preservation logic
  into a `usePreservedAgentUrl` hook so both items share it.
