+++ locales/en-US/chat.json
@@ -818,7 +818,7 @@
   "tool.intervention.viewParameters": "View parameters ({{count}})",
   "toolAuth.authorize": "Authorize",
   "toolAuth.authorizing": "Authorizing...",
-  "toolAuth.hint": "Without authorization or configuration, Skills may not work. This can limit the Agent or cause errors.",
+  "toolAuth.hint": "When Skills aren't authorized or configured, the related Skills won't work and the Agent's capabilities may be limited or run into errors.",
   "toolAuth.signIn": "Sign In",
   "toolAuth.title": "Authorize Skills for this Agent",
   "topic.checkOpenNewTopic": "Start a new topic?",
