+++ src/routes/(main)/home/_layout/Body/Agent/List/AgentItem/index.tsx
@@ -1,4 +1,3 @@
-import { SESSION_CHAT_URL } from '@lobechat/const';
 import { HETEROGENEOUS_TYPE_LABELS } from '@lobechat/heterogeneous-agents';
 import { type SidebarAgentItem } from '@lobechat/types';
 import { ActionIcon, Flexbox, Icon, Tag } from '@lobehub/ui';
@@ -7,7 +6,7 @@ import { Loader2, PinIcon } from 'lucide-react';
 import { type CSSProperties, type DragEvent } from 'react';
 import { memo, useCallback, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Link, useLocation } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 import NavItem from '@/features/NavPanel/components/NavItem';
 import { usePrefetchAgent } from '@/hooks/usePrefetchAgent';
@@ -18,12 +17,9 @@ import { useHomeStore } from '@/store/home';
 
 import { useAgentModal } from '../../ModalProvider';
 import Actions from '../Item/Actions';
+import { usePreservedAgentUrl } from '../usePreservedAgentUrl';
 import Avatar from './Avatar';
 import { useAgentDropdownMenu } from './useDropdownMenu';
-
-// Sub-routes that are agent-scoped views (not tied to a specific topic/task id),
-// safe to carry over when switching between agents from the sidebar switcher.
-const PRESERVED_AGENT_SUB_PATHS = new Set(['topics', 'profile', 'channel']);
 
 const styles = createStaticStyles(({ css, cssVar }) => ({
   badge: css`
@@ -119,19 +115,7 @@ const AgentItem = memo<AgentItemProps>(({ item, style, className, onNavigate }) 
     displayTitle
   );
 
-  // Get URL for this agent — when switching from within an agent's sub-view
-  // (e.g. /agent/A/topics), preserve the sub-route on the new agent so users
-  // don't lose their place. Only safe for agent-scoped views; topic/task ids
-  // belong to the previous agent and must not be carried over.
-  const { pathname } = useLocation();
-  const agentUrl = useMemo(() => {
-    const match = pathname.match(/^\/agent\/[^/]+\/([^/]+)\/?$/);
-    const subPath = match?.[1];
-    if (subPath && PRESERVED_AGENT_SUB_PATHS.has(subPath)) {
-      return `/agent/${id}/${subPath}`;
-    }
-    return SESSION_CHAT_URL(id, false);
-  }, [id, pathname]);
+  const agentUrl = usePreservedAgentUrl(id);
 
   // Memoize event handlers
   const handleMouseEnter = useCallback(() => {
